pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "document-annotation-service"
version = "0.1.0"
description = "Event-driven document annotation service take-home implementation"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.2",
  "fastapi>=0.111.0",
  "openpyxl>=3.1.5",
  "psycopg[binary]>=3.2.1",
  "pydantic-settings>=2.3.4",
  "pypdf>=5.1.0",
  "python-multipart>=0.0.9",
  "sqlalchemy>=2.0.31",
  "uvicorn[standard]>=0.30.1",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.2.2",
]

[tool.setuptools.packages.find]
include = ["app*"]
